Introduces node statuses
This will allow a single node (for example course, subject, video) to be in
one of several states (visible, invisible, trash). This will for example allow us
to delete an item without actually removing it, it will also allow us to a hide
an item, until it is ready to be published.

This commit adds statuses to all the nodes mentioned above. It does
however only implement delete functionality through the status for
courses, the rest are to follow in another commit.