Implements a new URL scheme in preparation for issue #46
The new URL scheme will require the year and semester in the URL, this
breaks backward compatibility.

This commit also introduces utility taglibs
sdu:createLinkToCourse/Subject/Video. These should make it easier to
link to different parts of the page. This also eliminates the "need" for
video/viewV.

The video index is now also exposed in the video model. This shouldn't
change the database scheme.