Upgrades from ChartJS version 1.0.2 to 2.2.1
This resolves issue #86. The problem was caused by ChartJS crashing due
to not being able to calculate the correct dimensions. This was most likely
caused by the dynamic rendering done in the dashboard. Quick testing
showed that this behavior was not present in the newest version of
ChartJS, hence the upgrade was made.